+++ src/entities/user/api/api.ts
@@ -1,5 +1,5 @@
 import { apiFetch } from "@shared/utils/apiFetch"
-import { User } from "../model/types"
+import { User, UserList } from "../model/types"
 
 export const getUsersList = async ({ limit = 0, select = "username,image" }) => {
   try {
@@ -19,5 +19,18 @@ export const getUserById = async (userId: number): Promise<User> => {
   } catch (error) {
     console.error("사용자 조회 오류:", error)
     throw new Error(`사용자 조회 오류: ${error}`)
+  }
+}
+
+export const getAllUsers = async (): Promise<User[]> => {
+  try {
+    const response = await apiFetch("/users?limit=0", {
+      method: "GET",
+      headers: { "Content-Type": "application/json" },
+    })
+    return response.users
+  } catch (error) {
+    console.error("사용자 목록 조회 오류:", error)
+    throw new Error(`사용자 목록 조회 오류: ${error}`)
   }
 }
