+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/index.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/entities/comment/api/api.ts","./src/entities/comment/model/types.ts","./src/entities/post/api/api.ts","./src/entities/post/model/type.ts","./src/entities/tag/api/api.ts","./src/entities/tag/model/model.ts","./src/entities/user/api/api.ts","./src/entities/user/lib/index.ts","./src/entities/user/model/types.ts","./src/entities/user/ui/userinfotext.tsx","./src/features/comment/api/api.ts","./src/features/comment/config/commentformvalue.ts","./src/features/comment/hooks/usecommentform.tsx","./src/features/comment/hooks/useselectedcomment.tsx","./src/features/comment/ui/commentinfobox.tsx","./src/features/comment/ui/createcommentmodal.tsx","./src/features/comment/ui/editcommentmodal.tsx","./src/features/filter/config/sortvalue.ts","./src/features/filter/model/types.ts","./src/features/filter/model/usesearchparams.ts","./src/features/filter/model/useurlsync.ts","./src/features/filter/utils/buildurlpath.ts","./src/features/modal/hooks/usemodal.ts","./src/features/post/api/api.ts","./src/features/post/config/pagavalue.ts","./src/features/post/hooks/usepostmodal.ts","./src/features/post/model/useselectedpost.ts","./src/features/post/ui/createpostmodal.tsx","./src/features/post/ui/editpostmodal.tsx","./src/features/post/ui/postdetailmodal.tsx","./src/features/post/ui/postlikeinfotablecell.tsx","./src/features/post/ui/postreactinfotablecell.tsx","./src/features/post/ui/postsearchfilters.tsx","./src/features/post/ui/posttitleinfotablecell.tsx","./src/features/post/ui/postuserinfotablecell.tsx","./src/features/user/api/api.ts","./src/features/user/config/authervalue.ts","./src/features/user/model/useselecteduser.ts","./src/features/user/ui/userdetailmodal.tsx","./src/pages/postsmanagerpage.tsx","./src/shared/model/type.ts","./src/shared/ui/index.tsx","./src/shared/ui/button/button.tsx","./src/shared/ui/dialog/dialog.tsx","./src/shared/ui/dialog/dialogcontent.tsx","./src/shared/ui/dialog/dialogheader.tsx","./src/shared/ui/dialog/dialogtitle.tsx","./src/shared/ui/dialog/index.ts","./src/shared/ui/input/input.tsx","./src/shared/ui/input/searchinput.tsx","./src/shared/ui/select/selectcontent.tsx","./src/shared/ui/select/selectitem.tsx","./src/shared/ui/select/selectroot.tsx","./src/shared/ui/select/selecttrigger.tsx","./src/shared/ui/table/table.tsx","./src/shared/ui/table/tablebody.tsx","./src/shared/ui/table/tablecell.tsx","./src/shared/ui/table/tablehead.tsx","./src/shared/ui/table/tableheader.tsx","./src/shared/ui/table/tablerow.tsx","./src/shared/ui/textarea/textarea.tsx","./src/shared/utils/highlighttext.tsx","./src/widgets/post/ui/postpagination.tsx","./src/widgets/post/ui/posttablebody.tsx","./src/widgets/post/ui/posttableheader.tsx","./src/widgets/post/ui/posttotaltable.tsx","./src/widgets/post/ui/postsheader.tsx","./src/widgets/post/ui/poststablecontainer.tsx","./src/widgets/post/ui/index.ts","./src/widgets/ui/footer.tsx","./src/widgets/ui/header.tsx"],"version":"5.8.3"}
+{"root":["./src/app.tsx","./src/index.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/entities/comment/api/api.ts","./src/entities/comment/model/types.ts","./src/entities/post/api/api.ts","./src/entities/post/model/type.ts","./src/entities/tag/api/api.ts","./src/entities/tag/model/model.ts","./src/entities/user/api/api.ts","./src/entities/user/lib/index.ts","./src/entities/user/model/types.ts","./src/entities/user/ui/userinfotext.tsx","./src/features/comment/api/api.ts","./src/features/comment/config/commentformvalue.ts","./src/features/comment/hooks/usecommentform.tsx","./src/features/comment/hooks/useselectedcomment.tsx","./src/features/comment/ui/commentinfobox.tsx","./src/features/comment/ui/createcommentmodal.tsx","./src/features/comment/ui/editcommentmodal.tsx","./src/features/filter/config/sortvalue.ts","./src/features/filter/model/types.ts","./src/features/filter/model/usesearchparams.ts","./src/features/filter/model/useurlsync.ts","./src/features/filter/utils/buildurlpath.ts","./src/features/modal/hooks/usemodal.ts","./src/features/post/api/api.ts","./src/features/post/config/pagavalue.ts","./src/features/post/hooks/usepostmodal.ts","./src/features/post/model/useselectedpost.ts","./src/features/post/ui/createpostmodal.tsx","./src/features/post/ui/editpostmodal.tsx","./src/features/post/ui/postdetailmodal.tsx","./src/features/post/ui/postlikeinfotablecell.tsx","./src/features/post/ui/postreactinfotablecell.tsx","./src/features/post/ui/postsearchfilters.tsx","./src/features/post/ui/posttitleinfotablecell.tsx","./src/features/post/ui/postuserinfotablecell.tsx","./src/features/user/api/api.ts","./src/features/user/config/authervalue.ts","./src/features/user/model/useselecteduser.ts","./src/features/user/ui/userdetailmodal.tsx","./src/pages/postsmanagerpage.tsx","./src/shared/model/type.ts","./src/shared/ui/index.tsx","./src/shared/ui/button/button.tsx","./src/shared/ui/dialog/dialog.tsx","./src/shared/ui/dialog/dialogcontent.tsx","./src/shared/ui/dialog/dialogheader.tsx","./src/shared/ui/dialog/dialogtitle.tsx","./src/shared/ui/dialog/index.ts","./src/shared/ui/input/input.tsx","./src/shared/ui/input/searchinput.tsx","./src/shared/ui/select/selectcontent.tsx","./src/shared/ui/select/selectitem.tsx","./src/shared/ui/select/selectroot.tsx","./src/shared/ui/select/selecttrigger.tsx","./src/shared/ui/table/table.tsx","./src/shared/ui/table/tablebody.tsx","./src/shared/ui/table/tablecell.tsx","./src/shared/ui/table/tablehead.tsx","./src/shared/ui/table/tableheader.tsx","./src/shared/ui/table/tablerow.tsx","./src/shared/ui/textarea/textarea.tsx","./src/shared/utils/apifetch.ts","./src/shared/utils/highlighttext.tsx","./src/widgets/post/ui/postpagination.tsx","./src/widgets/post/ui/posttablebody.tsx","./src/widgets/post/ui/posttableheader.tsx","./src/widgets/post/ui/posttotaltable.tsx","./src/widgets/post/ui/postsheader.tsx","./src/widgets/post/ui/poststablecontainer.tsx","./src/widgets/post/ui/index.ts","./src/widgets/ui/footer.tsx","./src/widgets/ui/header.tsx"],"version":"5.8.3"}
